split economy file and recruiting
