+++ shared/src/server/mod.rs
@@ -1,7 +1,6 @@
 pub mod ai;
 pub mod buildings;
 pub mod create_server;
-pub mod economy;
 pub mod entities;
 pub mod game_scenes;
 pub mod lobby;
